+++ packages/zwave-js/src/lib/driver/Driver.test.ts
@@ -6,6 +6,8 @@ import {
 	ZWaveErrorCodes,
 } from "@zwave-js/core";
 import { MessageHeaders, MockSerialPort } from "@zwave-js/serial";
+import { wait } from "alcalzone-shared/async";
+import { NodeStatus } from "zwave-js/src/lib/node/Types";
 import {
 	AssociationCCReport,
 	AssociationCommand,
@@ -24,7 +26,7 @@ import {
 import { MessageType } from "../message/Constants";
 import { Message, messageTypes } from "../message/Message";
 import { ZWaveNode } from "../node/Node";
-import { Driver } from "./Driver";
+import { Driver, ZWaveOptions } from "./Driver";
 
 // load the driver with stubbed out Serialport
 jest.mock("@zwave-js/serial", () => {
@@ -39,8 +41,11 @@ jest.mock("@zwave-js/serial", () => {
 
 const PORT_ADDRESS = "/tty/FAKE";
 
-async function createAndStartDriver() {
-	const driver = new Driver(PORT_ADDRESS, { skipInterview: true });
+async function createAndStartDriver(options: Partial<ZWaveOptions> = {}) {
+	const driver = new Driver(PORT_ADDRESS, {
+		...options,
+		skipInterview: true,
+	});
 	driver.on("error", () => {
 		/* swallow error events during testing */
 	});
@@ -55,6 +60,7 @@ async function createAndStartDriver() {
 	// Mock the value DB, because the original one will not be initialized
 	// with skipInterview: true
 	driver["_valueDB"] = new Map() as any;
+	driver["_valueDB"]!.close = () => Promise.resolve();
 
 	return {
 		driver,
@@ -333,19 +339,19 @@ describe("lib/driver/Driver => ", () => {
 			});
 
 			// Receive a CAN to trigger the resend check
-			expect(async () => {
-				await serialport.receiveData(req.serialize());
+			expect(() => {
+				serialport.receiveData(req.serialize());
 				jest.runAllTimers();
 			}).not.toThrow();
 		});
 
-		it("should correctly handle multiple messages in the receive buffer", async () => {
+		it("should correctly handle multiple messages in the receive buffer", () => {
 			// This buffer contains a SendData transmit report and a ManufacturerSpecific report
 			const data = Buffer.from(
 				"010700130f000002e6010e000400020872050086000200828e",
 				"hex",
 			);
-			await serialport.receiveData(data);
+			serialport.receiveData(data);
 
 			// Ensure the driver ACKed two messages
 			expect(serialport.writeStub).toBeCalledTimes(2);
@@ -753,6 +759,141 @@ describe("lib/driver/Driver => ", () => {
 				command: cc,
 			});
 			expect(() => driver["assemblePartialCCs"](msg)).toThrow("invalid");
+		});
+	});
+
+	describe("regression tests", () => {
+		let driver: Driver;
+		let serialport: MockSerialPort;
+		process.env.LOGLEVEL = "debug";
+
+		beforeEach(async () => {
+			({ driver, serialport } = await createAndStartDriver());
+
+			driver["_controller"] = {
+				ownNodeId: 1,
+				isFunctionSupported: () => true,
+				nodes: new Map(),
+			} as any;
+		});
+
+		afterEach(() => {
+			driver.destroy();
+			driver.removeAllListeners();
+		});
+
+		it("marking a node with a pending message as asleep does not mess up the remaining transactions", async () => {
+			jest.setTimeout(5000);
+			jest.useRealTimers();
+			// Repro from #1107
+
+			// Node 10's awake timer elapses before its ping is rejected,
+			// this causes mismatched responses for all following messages
+
+			const node10 = new ZWaveNode(10, driver);
+			const node17 = new ZWaveNode(17, driver);
+			(driver.controller.nodes as Map<number, ZWaveNode>).set(10, node10);
+			(driver.controller.nodes as Map<number, ZWaveNode>).set(17, node17);
+			// Add event handlers for the nodes
+			for (const node of driver.controller.nodes.values()) {
+				driver["addNodeEventHandlers"](node);
+			}
+
+			node10.addCC(CommandClasses["Wake Up"], { isSupported: true });
+			node10.markAsAwake();
+			expect(node10.status).toBe(NodeStatus.Awake);
+
+			driver["lastCallbackId"] = 2;
+			const ACK = Buffer.from([MessageHeaders.ACK]);
+
+			const pingPromise10 = node10.ping();
+			const pingPromise17 = node17.ping();
+			// » [Node 010] [REQ] [SendData]
+			//   │ transmit options: 0x25
+			//   │ callback id:      3
+			//   └─[NoOperationCC]
+			expect(serialport.lastWrite).toEqual(
+				Buffer.from("010800130a01002503c9", "hex"),
+			);
+			await wait(1);
+			serialport.receiveData(ACK);
+
+			await wait(15);
+
+			// « [RES] [SendData]
+			//     was sent: true
+			serialport.receiveData(Buffer.from("0104011301e8", "hex"));
+			// » [ACK]
+			expect(serialport.lastWrite).toEqual(ACK);
+
+			await wait(15);
+
+			node10.markAsAsleep();
+			expect(node10.status).toBe(NodeStatus.Asleep);
+
+			// The command queue should now abort the ongoing transaction
+			// » [REQ] [SendDataAbort]
+			expect(serialport.lastWrite).toEqual(
+				Buffer.from("01030016ea", "hex"),
+			);
+			await wait(1);
+			serialport.receiveData(ACK);
+
+			// Abort was acknowledged, Ping for 10 should be failed
+			await wait(15);
+			await expect(pingPromise10).resolves.toBe(false);
+
+			// Now the Ping for 17 should go out
+			// » [Node 017] [REQ] [SendData]
+			//   │ transmit options: 0x25
+			//   │ callback id:      4
+			//   └─[NoOperationCC]
+			expect(serialport.lastWrite).toEqual(
+				Buffer.from("010800131101002504d5", "hex"),
+			);
+			serialport.receiveData(ACK);
+
+			await wait(15);
+
+			// Callback for previous message comes and should be ignored
+			// « [REQ] [SendData]
+			//     callback id:     3
+			//     transmit status: NoAck
+			serialport.receiveData(
+				Buffer.from(
+					"011800130301019b007f7f7f7f7f010107000000000204000012",
+					"hex",
+				),
+			);
+			expect(serialport.lastWrite).toEqual(ACK);
+
+			await wait(1);
+
+			// Ping 17 does not get resolved by the other callback
+			await expect(Promise.race([pingPromise17, wait(15)])).resolves.toBe(
+				undefined,
+			);
+
+			// « [RES] [SendData]
+			//     was sent: true
+			serialport.receiveData(Buffer.from("0104011301e8", "hex"));
+			// » [ACK]
+			expect(serialport.lastWrite).toEqual(ACK);
+
+			await wait(15);
+
+			// Callback for ping node 17 (failed)
+			// « [REQ] [SendData]
+			//     callback id:     4
+			//     transmit status: NoAck
+			serialport.receiveData(
+				Buffer.from(
+					"011800130401019d007f7f7f7f7f010107000000000204000013",
+					"hex",
+				),
+			);
+			expect(serialport.lastWrite).toEqual(ACK);
+			await expect(pingPromise17).resolves.toBeFalse();
 		});
 	});
 });