+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * The type `CommandClass` is now exported from `zwave-js/CommandClass`
 * The interview process for `Configuration CC V3+` now continues even if the response to `NameGet` and/or `InfoGet` commands is missing or incomplete
 * Multi Channel Lifeline Associations are no longer created automatically if the device does not support the `Multi Channel CC`
+* Fixed an issue where marking nodes with active transaction as asleep would mess up the serial communication with the controller
 
 ## 5.3.6 (2020-11-04)
 ### Bugfixes